+++ HISTORY.rst
@@ -3,7 +3,10 @@ Upcoming
 
 New features:
 
-- ``solvcon.kerpak.cuse`` series modules are changed to use OpenMP for
+- Glue BCs are added.  A pair of collocated BCs can now be glued together to
+  work as an internal interface.  The glued BCs helps to dynamically turn on or
+  off the BC pair.
+- ``solvcon.kerpak.cuse`` series solver kernels are changed to use OpenMP for
   multi-threaded computing.  They were using a thread pool built-in SOLVCON for
   multi-threading.  OpenMP makes multi-threaded functions more flexible in
   argument specification.